Unit tests: made the TestEventModifier standalone in source file
